refactor(onboarding): separate content from types and helpers
